scripts/requirements.txt
jinja2
pillow
pyyaml
